fix(mobile): synchronize native appearance and composer layout
