+++ apps/mobile/src/components/shell/app-providers.tsx
@@ -1,3 +1,5 @@
+import { useResolvedColorScheme } from '@mobile/components/theme/use-color-scheme';
+import { DarkTheme, DefaultTheme, ThemeProvider } from 'expo-router';
 import { HeroUINativeProvider } from 'heroui-native';
 import { KeyboardProvider } from 'react-native-keyboard-controller';
 import { IntlProvider } from 'use-intl';
@@ -5,7 +7,12 @@ import { IntlProvider } from 'use-intl';
 /** Children-optional wrappers so the root layout can compose providers whose own
  * props declare `children` as required (mirrors workbench's AppI18nProvider pattern). */
 export function AppThemeProvider({ children }: React.PropsWithChildren): React.ReactNode {
-  return <HeroUINativeProvider>{children}</HeroUINativeProvider>;
+  const scheme = useResolvedColorScheme();
+  return (
+    <ThemeProvider value={scheme === 'dark' ? DarkTheme : DefaultTheme}>
+      <HeroUINativeProvider>{children}</HeroUINativeProvider>
+    </ThemeProvider>
+  );
 }
 
 /** Feeds keyboard frames to `KeyboardStickyView` so the composer tracks the keyboard. */
